chore: updated dependencies for attribute-popup and attribute-table
