+++ packages/attribute-popup/package.json
@@ -36,28 +36,31 @@
 	"homepage": "https://github.com/watergis/svelte-maplibre-components/tree/main/packages/attribute-popup#readme",
 	"devDependencies": {
 		"@arkweid/lefthook": "^0.7.7",
-		"@sveltejs/adapter-auto": "^3.1.1",
-		"@sveltejs/kit": "^2.5.4",
-		"@sveltejs/package": "^2.3.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.2",
-		"@typescript-eslint/eslint-plugin": "^7.2.0",
-		"@typescript-eslint/parser": "^7.2.0",
-		"eslint": "^8.57.0",
+		"@sveltejs/adapter-auto": "^3.2.4",
+		"@sveltejs/kit": "^2.5.24",
+		"@sveltejs/package": "^2.3.4",
+		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@typescript-eslint/eslint-plugin": "^8.2.0",
+		"@typescript-eslint/parser": "^8.2.0",
+		"eslint": "^9.9.1",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"prettier": "^3.2.5",
-		"prettier-plugin-svelte": "^3.2.2",
-		"sass": "^1.72.0",
-		"svelte": "^4.2.12",
-		"svelte-check": "^3.6.7",
-		"svelte-preprocess": "^5.1.3",
-		"tslib": "^2.6.2",
-		"typescript": "^5.4.2",
-		"vite": "^5.1.6"
+		"eslint-plugin-svelte": "^2.43.0",
+		"globals": "^15.9.0",
+		"pmtiles": "^3.0.5",
+		"prettier": "^3.3.3",
+		"prettier-plugin-svelte": "^3.2.6",
+		"sass": "^1.77.8",
+		"svelte": "^4.2.19",
+		"svelte-check": "^3.8.6",
+		"svelte-preprocess": "^6.0.2",
+		"tslib": "^2.7.0",
+		"typescript": "^5.5.4",
+		"typescript-eslint": "^8.2.0",
+		"vite": "^5.4.2"
 	},
 	"type": "module",
 	"dependencies": {
-		"maplibre-gl": "^4.1.0"
+		"maplibre-gl": "^4.6.0"
 	},
 	"peerDependencies": {
 		"svelte": "^4.0.0"
