+++ packages/attribute-popup/src/routes/+page.svelte
@@ -1,9 +1,57 @@
 <script lang="ts">
-	import Map from '$example/Map.svelte';
+	import AttributePopupControl from '$lib';
+	import { addProtocol, Map } from 'maplibre-gl';
+	import { Protocol } from 'pmtiles';
+	import { onMount } from 'svelte';
+
+	let mapContainer: HTMLDivElement;
+	let map: Map;
+
+	let targetLayers = [
+		'meter',
+		'flow meter',
+		'valve',
+		'washout',
+		'firehydrant',
+		'tank',
+		'pipeline',
+		'pg-building',
+		'sewer-connection',
+		'sewer_commercial',
+		'sewer_institution',
+		'sewer_public_toilet',
+		'sewer_pipeline',
+		'sewer_treatment_plant',
+		'manhole'
+	];
+
+	onMount(async () => {
+		const protocol = new Protocol();
+		addProtocol('pmtiles', protocol.tile);
+
+		map = new Map({
+			container: mapContainer,
+			style: 'https://narwassco.github.io/mapbox-stylefiles/unvt/style.json'
+		});
+	});
 </script>
 
 <sveltekit:head>
 	<title>Svelte maplibre legend example</title>
 </sveltekit:head>
 
-<Map />
+<div class="map" bind:this={mapContainer} />
+<AttributePopupControl bind:map {targetLayers} />
+
+<style>
+	@import 'maplibre-gl/dist/maplibre-gl.css';
+
+	.map {
+		position: absolute;
+		top: 0;
+		bottom: 0;
+		width: 100%;
+		height: 100%;
+		z-index: 1;
+	}
+</style>
