+++ packages/attribute-table/package.json
@@ -44,32 +44,34 @@
 		"svelte": "^4.0.0"
 	},
 	"devDependencies": {
-		"@sveltejs/adapter-auto": "^3.1.1",
-		"@sveltejs/kit": "^2.5.4",
-		"@sveltejs/package": "^2.3.0",
-		"@sveltejs/vite-plugin-svelte": "^3.0.2",
-		"@typescript-eslint/eslint-plugin": "^7.2.0",
-		"@typescript-eslint/parser": "^7.2.0",
-		"eslint": "^8.57.0",
+		"@sveltejs/adapter-auto": "^3.2.4",
+		"@sveltejs/kit": "^2.5.24",
+		"@sveltejs/package": "^2.3.4",
+		"@sveltejs/vite-plugin-svelte": "^3.1.2",
+		"@typescript-eslint/eslint-plugin": "^8.2.0",
+		"@typescript-eslint/parser": "^8.2.0",
+		"eslint": "^9.9.1",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
-		"pmtiles": "^3.0.5",
-		"prettier": "^3.2.5",
-		"prettier-plugin-svelte": "^3.2.2",
-		"publint": "^0.2.7",
-		"sass": "^1.72.0",
-		"svelte": "^4.2.12",
-		"svelte-check": "^3.6.7",
-		"tslib": "^2.6.2",
-		"typescript": "^5.4.2",
-		"vite": "^5.1.6"
+		"eslint-plugin-svelte": "^2.43.0",
+		"globals": "^15.9.0",
+		"pmtiles": "^3.0.7",
+		"prettier": "^3.3.3",
+		"prettier-plugin-svelte": "^3.2.6",
+		"publint": "^0.2.10",
+		"sass": "^1.77.8",
+		"svelte": "^4.2.19",
+		"svelte-check": "^3.8.6",
+		"tslib": "^2.7.0",
+		"typescript": "^5.5.4",
+		"typescript-eslint": "^8.2.0",
+		"vite": "^5.4.2"
 	},
 	"svelte": "./dist/index.js",
 	"types": "./dist/index.d.ts",
 	"type": "module",
 	"dependencies": {
-		"@vincjo/datatables": "^1.14.5",
+		"@vincjo/datatables": "^1.14.10",
 		"@watergis/svelte-maplibre-menu": "workspace:^",
-		"maplibre-gl": "^4.1.0"
+		"maplibre-gl": "^4.6.0"
 	}
 }
